balancer/requirements.txt
pyyaml
unidecode
pandas
gspread
python-dotenv